requirements.txt
deep-translator==1.11.4
flask==2.2.2
numpy==1.23.3
pandas==1.5.0
opencv-python==4.6.0.66
underthesea==1.3.5a3
pyvi==0.1.1
translate==3.6.1
faiss-cpu==1.7.2
ftfy==6.1.1
regex==2022.9.13
tqdm==4.64.1
matplotlib==3.6.0
langdetect==1.0.9
rapidfuzz==2.12.0
sentence_transformers==2.2.2
cloudinary==1.41.0
mysql-connector-python==9.0.0
ultralytics==8.2.79